Some refactoring done in the code: book share now have it's own controller (list and delete methods will be done yet)
